requirement.txt
numpy==1.18.5
scipy==1.9.3
scikit-learn==1.2.2
matplotlib==3.1.2
tensorflow==2.3.1
tensorflow-estimator==2.3.0
tensorboard==2.4.0
h5py==2.10.0
protobuf==3.19.4
grpcio==1.33.2
opt-einsum==3.3.0
absl-py==0.11.0
gast==0.3.3
google-pasta==0.2.0
wrapt==1.12.1
termcolor==1.1.0
six==1.14.0
typing-extensions==3.7.4.3
requests==2.22.0
urllib3==1.25.8
pillow==7.0.0
